Make: move ui.h to include directory
